+++ src/protocols/mod.rs
@@ -5,45 +5,37 @@
 use crate::connections::Connection;
 
 use once_cell::sync::OnceCell;
-use tokio::sync::{mpsc, oneshot};
+use tokio::sync::oneshot;
 
-use std::{io, net::SocketAddr};
+use std::io;
 
 mod disconnect;
 mod handshaking;
 mod reading;
 mod writing;
 
-pub use disconnect::Disconnect;
-pub use handshaking::Handshaking;
-pub use reading::Reading;
-pub use writing::Writing;
+pub use disconnect::{Disconnect, DisconnectHandler};
+pub use handshaking::{Handshaking, HandshakingHandler};
+pub use reading::{Reading, ReadingHandler};
+pub use writing::{Writing, WritingHandler};
 
 #[derive(Default)]
 pub(crate) struct Protocols {
-    pub(crate) handshake_handler: OnceCell<ProtocolHandler<ReturnableConnection>>,
-    pub(crate) reading_handler: OnceCell<ProtocolHandler<ReturnableConnection>>,
-    pub(crate) writing_handler: OnceCell<ProtocolHandler<ReturnableConnection>>,
-    pub(crate) disconnect_handler: OnceCell<ProtocolHandler<ReturnableItem<SocketAddr, ()>>>,
+    pub(crate) handshake_handler: OnceCell<HandshakingHandler>,
+    pub(crate) reading_handler: OnceCell<ReadingHandler>,
+    pub(crate) writing_handler: OnceCell<WritingHandler>,
+    pub(crate) disconnect_handler: OnceCell<DisconnectHandler>,
 }
 
-/// An object dedicated to managing a protocol; it contains a `Sender` whose other side is
-/// owned by the protocol's task.
-pub struct ProtocolHandler<T: Send>(mpsc::Sender<T>);
+/// The basic functionality that every protocol is required to handle - being triggered.
+#[async_trait::async_trait]
+pub trait ProtocolHandler {
+    /// The type of item that gets sent to the protocol handler task.
+    type Item;
 
-impl<T: Send> ProtocolHandler<T> {
-    /// Sends a returnable item to a task spawned by the protocol handler.
-    pub async fn send(&self, item: T) {
-        if self.0.send(item).await.is_err() {
-            unreachable!(); // protocol's task is down! can't recover
-        }
-    }
-}
-
-impl<T: Send> From<mpsc::Sender<T>> for ProtocolHandler<T> {
-    fn from(sender: mpsc::Sender<T>) -> Self {
-        Self(sender)
-    }
+    /// Triggers the protocol's behavior for the given subject, usually a `Connection` or
+    /// its `SocketAddr`.
+    async fn trigger(&self, item: Self::Item);
 }
 
 /// An object sent to a protocol handler task; the task assumes control of a protocol-relevant item `T`,